+++ backend/app/schemas/dag.py
@@ -21,8 +21,8 @@ class DagNodeCreate(BaseModel):
     comparison_variable: str | None = None
     default_prior_value: Decimal | None = None
     waterfall_order: int | None = None
-    position_x: int = 0
-    position_y: int = 0
+    position_x: float = 0
+    position_y: float = 0
 
 
 class DagEdgeCreate(BaseModel):
@@ -54,8 +54,8 @@ class DagNodeResponse(BaseModel):
     comparison_variable: str | None
     default_prior_value: Decimal | None
     waterfall_order: int | None
-    position_x: int
-    position_y: int
+    position_x: float
+    position_y: float
     is_active: bool
